requirements/base.txt
apolo-cli==24.7.1
apolo-sdk==24.7.1
click==8.1.2
pyyaml==6.0.1
toml==0.10.2
